--- a/.travis.yml
+++ b/.travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - "6"
+  - "8"
 before_script:
   - export DISPLAY=:99.0
   - sh -e /etc/init.d/xvfb start