--- a/src/index.ts
+++ b/src/index.ts
@@ -1,6 +1,7 @@
 import { Either, Left, Right, isRight } from 'fp-ts/lib/Either'
 
 import { Predicate } from 'fp-ts/lib/function'
+import { snoc } from 'fp-ts/lib/Array'
 
 declare global {
   interface Array<T> {
@@ -366,7 +367,7 @@ export const array = <RT extends Any>(
         const errors: Errors = []
         for (let i = 0; i < len; i++) {
           const x = xs[i]
-          const validation = type.validate(x, c.concat([getContextEntry(String(i), type)]))
+          const validation = type.validate(x, snoc(c)(getContextEntry(String(i), type)))
           validation.fold(
             e => pushAll(errors, e),
             vx => {
@@ -445,7 +446,7 @@ export const type = <P extends Props>(
         for (let k in props) {
           const ok = o[k]
           const type = props[k]
-          const validation = type.validate(ok, c.concat([getContextEntry(k, type)]))
+          const validation = type.validate(ok, snoc(c)(getContextEntry(k, type)))
           validation.fold(
             e => pushAll(errors, e),
             vok => {
@@ -554,8 +555,8 @@ export const dictionary = <D extends Any, C extends Any>(
         let changed = false
         for (let k in o) {
           const ok = o[k]
-          const domainValidation = domain.validate(k, c.concat([getContextEntry(k, domain)]))
-          const codomainValidation = codomain.validate(ok, c.concat([getContextEntry(k, codomain)]))
+          const domainValidation = domain.validate(k, snoc(c)(getContextEntry(k, domain)))
+          const codomainValidation = codomain.validate(ok, snoc(c)(getContextEntry(k, codomain)))
           domainValidation.fold(
             e => pushAll(errors, e),
             vk => {
@@ -615,7 +616,7 @@ export const union = <RTS extends Array<Any>>(
       const errors: Errors = []
       for (let i = 0; i < len; i++) {
         const type = types[i]
-        const validation = type.validate(s, c.concat([getContextEntry(String(i), type)]))
+        const validation = type.validate(s, snoc(c)(getContextEntry(String(i), type)))
         if (isRight(validation)) {
           return validation
         } else {
@@ -743,7 +744,7 @@ export function tuple<RTS extends Array<Any>>(
         for (let i = 0; i < len; i++) {
           const a = as[i]
           const type = types[i]
-          const validation = type.validate(a, c.concat([getContextEntry(String(i), type)]))
+          const validation = type.validate(a, snoc(c)(getContextEntry(String(i), type)))
           validation.fold(
             e => pushAll(errors, e),
             va => {
@@ -757,7 +758,7 @@ export function tuple<RTS extends Array<Any>>(
           )
         }
         if (as.length > len) {
-          errors.push(getValidationError(as[len], c.concat([getContextEntry(String(len), never)])))
+          errors.push(getValidationError(as[len], snoc(c)(getContextEntry(String(len), never))))
         }
         return errors.length ? failures(errors) : success(t as any)
       }),
@@ -876,7 +877,7 @@ export const strict = <P extends Props>(
           for (let i = 0; i < keys.length; i++) {
             const key = keys[i]
             if (!props.hasOwnProperty(key)) {
-              errors.push(getValidationError(o[key], c.concat([getContextEntry(key, never)])))
+              errors.push(getValidationError(o[key], snoc(c)(getContextEntry(key, never))))
             }
           }
           return errors.length ? failures(errors) : failure(o, c)