--- a/test/perf.ts
+++ b/test/perf.ts
@@ -1,9 +1,20 @@
-import * as assert from 'assert'
 import * as Benchmark from 'benchmark'
-import { Context, getDefaultContext, getContextEntry, string, number } from '../src/index'
+import * as assert from 'assert'
+
+import { Context, getContextEntry, getDefaultContext, number, string } from '../src/index'
 
 const suite = new Benchmark.Suite()
 const c: Context = getDefaultContext(string)
+
+const appendNext = (arr: any[], e: any) => {
+  const len = arr.length
+  const res = Array(len + 1)
+  for (let i = 0; i < len; i++) {
+    res[i] = arr[i]
+  }
+  res[len] = e
+  return res
+}
 
 describe('perf', () => {
   ;(it('concat([a]) should be faster than concat(a)', done => {
@@ -13,6 +24,9 @@ describe('perf', () => {
       })
       .add('concat([e])', function() {
         c.concat([getContextEntry('key', number)])
+      })
+      .add('manual appendNext function', function() {
+        appendNext(c, getContextEntry('key', number))
       })
       .on('cycle', function(event: any) {
         console.log(String(event.target))