--- a/test/tsconfig.json
+++ b/test/tsconfig.json
@@ -13,7 +13,6 @@
     "allowJs": false,
     "target": "es5",
     "moduleResolution": "node",
-    "forceConsistentCasingInFileNames": true,
-    "lib": ["es6", "dom"]
+    "forceConsistentCasingInFileNames": true
   }
 }