Using `./perf/index.js`

---

`concat(element)` (current implementation)

invalid payload x 200,425 ops/sec ±0.74% (89 runs sampled)
valid payload (no errors) x 181,576 ops/sec ±1.00% (87 runs sampled)

`concat([element])`

invalid payload x 258,902 ops/sec ±0.61% (89 runs sampled)
valid payload (no errors) x 227,362 ops/sec ±1.54% (86 runs sampled)