@gcanti 2017-12-12
It's pretty surprising by the way, I'm doing something wrong?

This is another benchmark

```ts
var Benchmark = require('benchmark')

const suite = new Benchmark.Suite()

const x = [1, 2, 3]

suite
  .add('concat(a)', function() {
    x.concat(4)
  })
  .add('concat([a])', function() {
    x.concat([4])
  })
  .on('cycle', function(event) {
    console.log(String(event.target))
  })
  .on('complete', function() {
    console.log('Fastest is ' + this.filter('fastest').map('name'))
  })
  .run({ async: true })
```

Results

```
x.concat(4) // 2,935,475 ops/sec ±0.87% (82 runs sampled)
x.concat([4]) // 5,421,103 ops/sec ±0.93% (86 runs sampled)
```