@sledorze 2017-12-12
@gcanti that's crazy, maybe due to recent optimisations changes in V8.
I'm wondering if, as we decide to choose an implementation over another based on a comparison, if we should not enforce in a test that, so that if it ever change, we can adapt..

`x.concat([4])` must be faster than `x.concat(4)`