@gcanti 2017-12-12
> if we should not enforce in a test that, so that if it ever change, we can adapt..

That would be nice